[feature] Mongo DB databases inventory
This commit simplify the databases inventory and make it generic.

Entries like:  
```yaml
database:
  user: someuser
  password: somepassword
```
will be parsed to create user and db as inventoried.

Also, the filenames in the Keybase directory are now
`mongodb_db_inventory-test.yml` and `mongodb_db_inventory-prod.yml`, 
where the environment suffix is taken from the `inventory_environment`
variable.